bootloaders: uint32_t -> uintptr_t

+++ flight/targets/Bootloaders/CopterControl/pios_board.c
@@ -79,7 +79,7 @@ void PIOS_Board_Init(void) {
 	/* Activate the HID-only USB configuration */
 	PIOS_USB_DESC_HID_ONLY_Init();
 
-	uint32_t pios_usb_id;
+	uintptr_t pios_usb_id;
 	switch(bdinfo->board_rev) {
 		case BOARD_REVISION_CC:
 			PIOS_USB_Init(&pios_usb_id, &pios_usb_main_cfg_cc);
@@ -91,7 +91,7 @@ void PIOS_Board_Init(void) {
 			PIOS_Assert(0);
 	}
 #if defined(PIOS_INCLUDE_USB_HID) && defined(PIOS_INCLUDE_COM_MSG)
-	uint32_t pios_usb_hid_id;
+	uintptr_t pios_usb_hid_id;
 	if (PIOS_USB_HID_Init(&pios_usb_hid_id, &pios_usb_hid_cfg, pios_usb_id)) {
 		PIOS_Assert(0);
 	}
